fixed issue with boxed primitive classes. need to fix overrides detection

+++ test/src/main/java/org/projectparams/test/Abobus.java
@@ -5,7 +5,7 @@ import org.projectparams.annotations.DefaultValue;
 public class Abobus {
     public Abobus(@DefaultValue("3.4") float abobus) {
     }
-    public float bibus(@DefaultValue("3.4") float abobus) {
+    public float bibus(@DefaultValue("3.4") Double abobus) {
         return this.doubleBibus(new Main());
     }
     public float doubleBibus(Main main, @DefaultValue("3.4") float abobus2) {
@@ -33,7 +33,9 @@ public class Abobus {
         public Bibus(@DefaultValue("3.4") float abobus) {
             super(abobus);
         }
-        public Bibus ultima(@DefaultValue("3.4") float abobus, @DefaultValue("true") boolean abobus2, @DefaultValue("4") int abobus3) {
+        public Bibus ultima(@DefaultValue("3.4") float abobus,
+                            @DefaultValue("true") boolean abobus2,
+                            @DefaultValue("4") int abobus3) {
             return new Bibus(abobus);
         }
 //        public UltimaPopus ultima(@DefaultValue("3.4") float abobus) {
@@ -41,7 +43,7 @@ public class Abobus {
 //        }
 
         public class UltimaPopus extends Bibus {
-            public UltimaPopus(Abobus abobus45, @DefaultValue("3.4") float abobus, @DefaultValue("5") byte abobus2) {
+            public UltimaPopus(Abobus abobus45, @DefaultValue("3.4") float abobus, @DefaultValue("5") Byte abobus2) {
                 super(abobus);
             }
             public UltimaPopus ororos(@DefaultValue("3.4") float abobus) {
